Added video uploader to success page and shortcut icon
If you say "yes" to the question of if you fulfilled your commitment, it
redirects you to a form where you can upload a video.  At this point it
still just does a JavaScript alert with the token because I couldn't
figure out yet how to make that value to into the hidden field on the
form.
Also, if you make a shortcut to the website on iOS or Android it should
have the red rectangular rails logo as the shortcut image (waiting on
upload of Caitlin's logo)